+++ PresTrustFloodMitigationService/Src/PresTrust.FloodMitigation.Domain/Entities/FloodTechDetailsEntity.cs
@@ -33,10 +33,10 @@ public class FloodTechDetailsEntity
     public string? ElevationBeforeMitigation { get; set; }
     public string? ElevationBeforeMitigationFinal { get; set; }
     public string? FloodType { get; set; }
-    public int? TenPercent { get; set; }
-    public int? TwoPercent { get; set; }
-    public int? OnePercent { get; set; }
-    public int? PointOnePercent { get; set; }
+    public decimal? TenPercent { get; set; }
+    public decimal? TwoPercent { get; set; }
+    public decimal? OnePercent { get; set; }
+    public decimal? PointOnePercent { get; set; }
     public string LastUpdatedBy { get; set; }
     public DateTime? LastUpdatedOn { get; set; }
 }
